pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dstdsv"
dynamic = ["version"]
description = 'Python driver for IMADA DST/DSV Force gauges'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Florian Dupeyron", email = "florian.dupeyron@mugcat.fr" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
	"pyserial"
]

[project.urls]
#Documentation = "https://github.com/U.N. Owen/dstdsv#readme"
#Issues = "https://github.com/U.N. Owen/dstdsv/issues"
#Source = "https://github.com/U.N. Owen/dstdsv"

[tool.hatch.version]
path = "src/dstdsv/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dstdsv tests}"

[tool.coverage.run]
source_pkgs = ["dstdsv", "tests"]
branch = true
parallel = true
omit = [
  "src/dstdsv/__about__.py",
]

[tool.coverage.paths]
dstdsv = ["src/dstdsv", "*/dstdsv/src/dstdsv"]
tests = ["tests", "*/dstdsv/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
